Added isotropic turbulence test with top/bottom walls.
This test seeds the velocity at all scales with the given power
spectrum and then normalizes the velocity to amp0.  The random
seed is compactly represented as a sum over log_2(N) random
velocity vectors which are flipped based on the bitstring
representation of the point's integer index.  This ensures
that the total velocity is zero.

I am not sure if this really qualifies as isotropic turbulence, but
its a start at least...